feat(de schema): use name+shortName max length from schema
