+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2023-11-16T09:08:17.016Z\n"
-"PO-Revision-Date: 2023-11-16T09:08:17.016Z\n"
+"POT-Creation-Date: 2023-11-16T09:15:47.674Z\n"
+"PO-Revision-Date: 2023-11-16T09:15:47.674Z\n"
 
 msgid "schemas"
 msgstr "schemas"
@@ -760,6 +760,9 @@ msgstr "Option set comment"
 
 msgid "Choose a set of predefined comment for data entry"
 msgstr "Choose a set of predefined comment for data entry"
+
+msgid "This {{field}} already exists, please choose antoher one"
+msgstr "This {{field}} already exists, please choose antoher one"
 
 msgid "Metadata management"
 msgstr "Metadata management"
