+++ src/lib/index.ts
@@ -2,7 +2,8 @@ export * from './constants'
 export * from './debounce'
 export * from './models'
 export * from './schemas'
-export * from './useLoadApp'
+export { useLoadApp } from './useLoadApp'
+export type { ModelSchemas, Schema } from './useLoadApp'
 export * from './errors'
 export * from './user'
 export * from './sections'
