+++ src/pages/dataElements/form/index.ts
@@ -2,4 +2,4 @@ export { DataElementFormFields } from './DataElementFormFields'
 export { computeInitialValues } from './computeInitialValues'
 export { useCustomAttributesQuery } from './useCustomAttributesQuery'
 export type { FormValues } from './types'
-export { validate } from './validate'
+export { useValidate } from './useValidate'
